Enable L7 redirection based on X-FORWARDED-PROTO
This commit lets users specify a `x_forwarded_proto_redirect`
field in their ambassador mapping which accepts a boolen value.

An example mapping looks like -

```yaml
    getambassador.io/config: |
      ---
      apiVersion: ambassador/v0
      kind:  Mapping
      name:  httpbin_mapping
      prefix: /httpbin/
      service: httpbin.org:80
      host_rewrite: httpbin.org
      x_forwarded_proto_redirect: true
```

When set to true, ambassador, and in turn envoy, behaves like the
following -

If an incoming request has `X-FORWARDED-PROTO: https` header
set, then no redirects take place and a normal response if sent.

If `X-FORWARDED-PROTO` is absent or sent to anything else, like
`http`, then a 301 redirect to https://host is sent as response.

Keep in mind that ambassador does not have access to any TLS certs
in the entire workflow. This is typically useful in cases when TLS
termination happens at an external L7 load balancer, which sets
the protocol of the incoming request in the `X-FORWARDED-PROTO`
header of the resulting HTTP request after TLS termination.

Partially fixes #469